requirements.txt
open3d>=0.10.0.0
numpy>=1.18.5
pandas>=1.1.2
matplotlib>=3.2.1
scikit-image>=0.17.2
tensorflow>=2.3.1
scipy>=1.5.4
sklearn>=0.0
scikit-learn~=0.23.2
tqdm>=4.55.1
ipython>=7.15.0
Pillow>=7.2.0